.llm/case-02-telegram/requirements.txt
anthropic
python-telegram-bot==20.7
sqlalchemy
psycopg2-binary
pandas
tabulate
python-dotenv
pyyaml
